requirements.txt
# Enterprise Trading Bot Requirements - Python 3.12 Compatible
# =============================================================

# Core Framework
fastapi==0.114.0
uvicorn[standard]==0.25.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Async HTTP
aiohttp==3.9.5

# Data Analysis & Trading  
pandas==2.2.0
numpy==1.26.4

# Environment & Config
python-dotenv==1.0.1

# Logging & Monitoring
structlog==24.1.0

# WebSocket
websockets==12.0

# Timezone
pytz==2024.1

# Crypto & Security (for API signatures)
cryptography==42.0.5

# Performance & Caching
cachetools==5.3.3
aiolimiter==1.2.1
bingx==0.0.73
backoff==2.2.1
fastapi-cache2==0.2.2

# Development & Testing (optional)
pytest==8.0.0
pytest-asyncio==0.23.5

# Python 3.12 Compatibility
setuptools==69.1.0